Analyser filter (#97)
* refactor filters

* clean up switch case

* better CSS for filter

* prettify everything
